requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
alembic==1.16.5
pydantic==2.11.7
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pandas==2.3.2
numpy==2.0.2
scipy==1.15.3
statsmodels==0.14.5
plotly==6.3.0
jinja2==3.1.6
weasyprint==66.0
streamlit==1.49.1
requests==2.32.5
pytest==8.4.2
httpx==0.28.1
